chore: move page-card components to separate folder
