+++ packages/apps/human-app/frontend/src/modules/operator/views/sign-up/add-keys.page.tsx
@@ -2,9 +2,9 @@ import { Grid } from '@mui/material';
 import type { UseFormReturn } from 'react-hook-form';
 import { t } from 'i18next';
 import { Link } from 'react-router-dom';
-import { PageCardError } from '@/shared/components/ui/page-card-error';
-import { PageCardLoader } from '@/shared/components/ui/page-card-loader';
-import { PageCard } from '@/shared/components/ui/page-card';
+import { PageCardError } from '@/shared/components/ui/page-card/page-card-error';
+import { PageCardLoader } from '@/shared/components/ui/page-card/page-card-loader';
+import { PageCard } from '@/shared/components/ui/page-card/page-card';
 import { getErrorMessageForError, jsonRpcErrorHandler } from '@/shared/errors';
 import type { EditEthKVStoreValuesMutationData } from '@/modules/operator/hooks/use-edit-existing-keys';
 import { useEditExistingKeysMutationState } from '@/modules/operator/hooks/use-edit-existing-keys';
