+++ packages/apps/human-app/frontend/src/modules/worker/views/email-verification/verify-email.page.tsx
@@ -5,8 +5,8 @@ import { useNavigate } from 'react-router-dom';
 import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { FormProvider, useForm } from 'react-hook-form';
-import { PageCard } from '@/shared/components/ui/page-card';
-import { PageCardLoader } from '@/shared/components/ui/page-card-loader';
+import { PageCard } from '@/shared/components/ui/page-card/page-card';
+import { PageCardLoader } from '@/shared/components/ui/page-card/page-card-loader';
 import { useLocationState } from '@/modules/worker/hooks/use-location-state';
 import { env } from '@/shared/env';
 import type { ResendEmailVerificationDto } from '@/modules/worker/services/resend-email-verification';
