+++ packages/apps/human-app/frontend/src/modules/worker/views/hcaptcha-labeling/hcaptcha-labeling.page.tsx
@@ -9,8 +9,8 @@ import { env } from '@/shared/env';
 import { breakpoints } from '@/shared/styles/breakpoints';
 import { Counter } from '@/shared/components/ui/counter';
 import { useHCaptchaUserStats } from '@/modules/worker/services/hcaptcha-user-stats';
-import { PageCardError } from '@/shared/components/ui/page-card-error';
-import { PageCardLoader } from '@/shared/components/ui/page-card-loader';
+import { PageCardError } from '@/shared/components/ui/page-card/page-card-error';
+import { PageCardLoader } from '@/shared/components/ui/page-card/page-card-loader';
 import { getErrorMessageForError } from '@/shared/errors';
 import { useDailyHmtSpent } from '@/modules/worker/services/daily-hmt-spent';
 import { getTomorrowDate } from '@/shared/helpers/date';
